# apidog.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Dragon Lap MVP Backend API",
    "description": "Single-tenant laptop store backend (NestJS + Prisma + JWT).",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development"
    }
  ],
...
  "security": [
    {
      "cookieAuth": []
    }
  ],
  "paths": {
...
    "/laptops": {
      "get": {
        "tags": ["Laptops"],
        "summary": "List published laptops",
        "security": [],
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "brand", "in": "query", "schema": { "type": "string" } },
          {
            "name": "categoryId",
            "in": "query",
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "minPrice",
            "in": "query",
            "schema": { "type": "number", "minimum": 0 }
          },
          {
            "name": "maxPrice",
            "in": "query",
            "schema": { "type": "number", "minimum": 0 }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["price_asc", "price_desc", "newest", "oldest"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Laptops",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Laptop" }
                }
              }
            }
          }
        }
      }
    },
    "/laptops/{id}": {
      "get": {
        "tags": ["Laptops"],
        "summary": "Get published laptop by ID",
        "security": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Laptop details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Laptop" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/admin/laptops": {
      "get": {
        "tags": ["Admin", "Laptops"],
        "summary": "List all laptops (admin)",
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          { "name": "brand", "in": "query", "schema": { "type": "string" } },
          {
            "name": "categoryId",
            "in": "query",
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["price_asc", "price_desc", "newest", "oldest"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All laptops",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Laptop" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      },
      "post": {
        "tags": ["Admin", "Laptops"],
        "summary": "Create laptop (admin)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateLaptopDto" },
              "example": {
                "title": "Lenovo Legion Pro 5",
                "brand": "Lenovo",
                "model": "16IRX9",
                "price": 76999,
                "stock": 8,
                "cpu": "Intel Core i7-14650HX",
                "ram": "16GB",
                "storage": "1TB NVMe",
                "gpu": "RTX 4060",
                "screenSize": "16 inch",
                "os": "Windows 11",
                "categoryId": "be54f9a8-6f99-4bb3-9f24-93e592c4493e",
                "shortDescription": "High-performance gaming laptop",
                "description": "QHD+ display, high refresh rate, RTX graphics",
                "isPublished": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Laptop" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/admin/laptops/{id}": {
      "patch": {
        "tags": ["Admin", "Laptops"],
        "summary": "Update laptop (admin)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateLaptopDto" },
              "example": {
                "price": 74999,
                "stock": 12,
                "isPublished": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Laptop" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Admin", "Laptops"],
        "summary": "Soft delete (unpublish) laptop (admin)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Soft deleted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Laptop" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/admin/laptops/{id}/images/sign": {
      "post": {
        "tags": ["Admin", "Laptops"],
        "summary": "Generate Cloudinary signed upload URLs (admin)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SignLaptopImagesDto" },
              "example": {
                "slots": ["featured", "gallery_1"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Signed upload parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignedUploadResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/admin/laptops/{id}/images": {
      "patch": {
        "tags": ["Admin", "Laptops"],
        "summary": "Attach uploaded images to laptop (admin)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SetLaptopImagesDto" },
              "example": {
                "featured": {
                  "url": "https://res.cloudinary.com/your-cloud/image/upload/v1234567890/dragon-lap/laptops/uuid/featured.jpg",
                  "publicId": "dragon-lap/laptops/uuid/featured"
                },
                "gallery": [
                  {
                    "url": "https://res.cloudinary.com/your-cloud/image/upload/v1234567890/dragon-lap/laptops/uuid/gallery_1.jpg",
                    "publicId": "dragon-lap/laptops/uuid/gallery_1"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Images attached",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Laptop" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/orders": {
      "post": {
        "tags": ["Orders"],
        "summary": "Place COD order from current cart",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateOrderDto" },
              "example": {
                "fullName": "Ahmed Ali",
                "phone": "01011223344",
                "alternatePhone": "01099887766",
                "city": "Cairo",
                "area": "Nasr City",
                "streetAddress": "Building 20, Street 10",
                "notes": "Call before delivery"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order placed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OrderWithItems" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/orders/guest": {
      "post": {
        "tags": ["Orders"],
        "summary": "Place guest COD order from guest cart",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateOrderDto" },
              "example": {
                "fullName": "Guest User",
                "phone": "01011223344",
                "alternatePhone": "01099887766",
                "city": "Cairo",
                "area": "Nasr City",
                "streetAddress": "Building 20, Street 10",
                "notes": "Call before delivery"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Guest order placed",
            "headers": {
              "Set-Cookie": {
                "description": "guest_session_id HTTP-only cookie is set if missing",
                "schema": { "type": "string" }
              }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OrderWithItems" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/orders/my": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get current customer orders",
        "responses": {
          "200": {
            "description": "My orders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/OrderWithItems" }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/orders/my/{id}": {
      "get": {
        "tags": ["Orders"],
        "summary": "Get one customer order by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Order",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OrderWithItems" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/admin/orders": {
      "get": {
        "tags": ["Admin", "Orders"],
        "summary": "List all orders (admin)",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/OrderStatusEnum"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Orders list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderWithItemsAndUser"
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/admin/orders/{id}": {
      "get": {
        "tags": ["Admin", "Orders"],
        "summary": "Get one order (admin)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Order details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderWithItemsAndUser"
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/admin/orders/{id}/status": {
      "patch": {
        "tags": ["Admin", "Orders"],
        "summary": "Update order status with transition rules (admin)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateOrderStatusDto" },
              "example": {
                "status": "CONFIRMED"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Order" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/admin/stats": {
      "get": {
        "tags": ["Admin"],
        "summary": "Get store statistics and recent activity (admin only)",
        "responses": {
          "200": {
            "description": "Store statistics",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AdminStats" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token",
        "description": "HTTP-only JWT cookie set automatically on login/register. Send requests with credentials (withCredentials: true)."
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Validation or business rule failure",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "statusCode": 400,
              "message": "Requested quantity exceeds stock",
              "error": "Bad Request"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid token",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "statusCode": 401,
              "message": "Unauthorized",
              "error": "Unauthorized"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Role/ownership denied",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "statusCode": 403,
              "message": "Insufficient role",
              "error": "Forbidden"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": {
              "statusCode": 404,
              "message": "Not found",
              "error": "Not Found"
            }
          }
        }
      }
    },
    "schemas": {
...
      "PaymentMethod": {
        "type": "string",
        "enum": ["COD"]
      },
      "OrderStatusEnum": {
        "type": "string",
        "enum": [
          "PENDING",
          "CONFIRMED",
          "PREPARING",
          "OUT_FOR_DELIVERY",
          "DELIVERED",
          "CANCELLED"
        ]
      },
...
      "CreateLaptopDto": {
        "type": "object",
        "required": ["title", "brand", "model", "price", "stock"],
        "properties": {
          "title": { "type": "string" },
          "slug": { "type": "string" },
          "brand": { "type": "string" },
          "model": { "type": "string" },
          "price": { "type": "number", "minimum": 0 },
          "stock": { "type": "number", "minimum": 0 },
          "shortDescription": { "type": "string" },
          "description": { "type": "string" },
          "cpu": { "type": "string" },
          "ram": { "type": "string" },
          "storage": { "type": "string" },
          "gpu": { "type": "string" },
          "screenSize": { "type": "string" },
          "os": { "type": "string" },
          "categoryId": { "type": "string", "format": "uuid" },
          "isPublished": { "type": "boolean" }
        }
      },
      "UpdateLaptopDto": {
        "allOf": [{ "$ref": "#/components/schemas/CreateLaptopDto" }]
      },
...
      "CreateOrderDto": {
        "type": "object",
        "required": ["fullName", "phone", "city", "area", "streetAddress"],
        "properties": {
          "fullName": { "type": "string" },
          "phone": { "type": "string" },
          "alternatePhone": { "type": "string" },
          "city": { "type": "string" },
          "area": { "type": "string" },
          "streetAddress": { "type": "string" },
          "notes": { "type": "string" }
        }
      },
      "UpdateOrderStatusDto": {
        "type": "object",
        "required": ["status"],
        "properties": {
          "status": { "$ref": "#/components/schemas/OrderStatusEnum" }
        }
      },
      "SignLaptopImagesDto": {
        "type": "object",
        "properties": {
          "slots": {
            "type": "array",
            "minItems": 1,
            "maxItems": 4,
            "items": {
              "type": "string",
              "enum": ["featured", "gallery_1", "gallery_2", "gallery_3"]
            },
            "description": "Slot names to generate signatures for. If omitted, all slots are signed."
          }
        }
      },
      "CloudinarySignature": {
        "type": "object",
        "properties": {
          "slot": { "type": "string" },
          "uploadUrl": { "type": "string", "format": "uri" },
          "cloudName": { "type": "string" },
          "apiKey": { "type": "string" },
          "folder": { "type": "string" },
          "publicId": { "type": "string" },
          "timestamp": { "type": "integer" },
          "signature": { "type": "string" },
          "description": "Use these values to construct FormData for direct Cloudinary upload"
        }
      },
      "SignedUploadResponse": {
        "type": "object",
        "properties": {
          "laptopId": { "type": "string", "format": "uuid" },
          "signatures": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CloudinarySignature" }
          }
        }
      },
      "LaptopImageInputDto": {
        "type": "object",
        "required": ["url", "publicId"],
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Cloudinary image URL after upload"
          },
          "publicId": {
            "type": "string",
            "description": "Cloudinary public ID of uploaded image"
          }
        }
      },
      "SetLaptopImagesDto": {
        "type": "object",
        "required": ["featured", "gallery"],
        "properties": {
          "featured": {
            "$ref": "#/components/schemas/LaptopImageInputDto"
          },
          "gallery": {
            "type": "array",
            "maxItems": 3,
            "items": { "$ref": "#/components/schemas/LaptopImageInputDto" },
            "description": "Up to 3 gallery images"
          }
        }
      },
...
      "LaptopImageKind": {
        "type": "string",
        "enum": ["FEATURED", "GALLERY"]
      },
      "LaptopImage": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "kind": { "$ref": "#/components/schemas/LaptopImageKind" },
          "url": { "type": "string", "format": "uri" },
          "publicId": { "type": "string" },
          "position": { "type": "integer" }
        }
      },
      "Laptop": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "slug": { "type": "string" },
          "brand": { "type": "string" },
          "model": { "type": "string" },
          "price": { "type": "string", "example": "76999.00" },
          "stock": { "type": "integer" },
          "shortDescription": { "type": "string", "nullable": true },
          "description": { "type": "string", "nullable": true },
          "cpu": { "type": "string", "nullable": true },
          "ram": { "type": "string", "nullable": true },
          "storage": { "type": "string", "nullable": true },
          "gpu": { "type": "string", "nullable": true },
          "screenSize": { "type": "string", "nullable": true },
          "os": { "type": "string", "nullable": true },
          "categoryId": { "type": "string", "format": "uuid", "nullable": true },
          "category": {
            "oneOf": [
              { "$ref": "#/components/schemas/Category" },
              { "type": "null" }
            ]
          },
          "featuredImage": {
            "oneOf": [
              { "$ref": "#/components/schemas/LaptopImage" },
              { "type": "null" }
            ]
          },
          "galleryImages": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/LaptopImage" }
          },
          "isPublished": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Category": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "OrderOwnerType": {
        "type": "string",
        "enum": ["CUSTOMER", "GUEST"]
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "userId": { "type": "string", "format": "uuid", "nullable": true },
          "ownerType": { "$ref": "#/components/schemas/OrderOwnerType" },
          "guestSessionId": { "type": "string", "nullable": true },
          "status": { "$ref": "#/components/schemas/OrderStatusEnum" },
          "paymentMethod": { "$ref": "#/components/schemas/PaymentMethod" },
          "fullName": { "type": "string" },
          "phone": { "type": "string" },
          "alternatePhone": { "type": "string", "nullable": true },
          "city": { "type": "string" },
          "area": { "type": "string" },
          "streetAddress": { "type": "string" },
          "notes": { "type": "string", "nullable": true },
          "subtotal": { "type": "string", "example": "153998.00" },
          "shippingFee": { "type": "string", "example": "0.00" },
          "total": { "type": "string", "example": "153998.00" },
          "placedAt": { "type": "string", "format": "date-time" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "orderId": { "type": "string", "format": "uuid" },
          "laptopId": { "type": "string", "format": "uuid", "nullable": true },
          "laptopTitleSnapshot": { "type": "string" },
          "unitPriceSnapshot": { "type": "string", "example": "76999.00" },
          "quantity": { "type": "integer" },
          "lineTotal": { "type": "string", "example": "153998.00" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "OrderWithItems": {
        "allOf": [
          { "$ref": "#/components/schemas/Order" },
          {
            "type": "object",
            "properties": {
              "items": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/OrderItem" }
              }
            }
          }
        ]
      },
      "OrderWithItemsAndUser": {
        "allOf": [
          { "$ref": "#/components/schemas/OrderWithItems" },
          {
            "type": "object",
            "properties": {
              "user": {
                "oneOf": [
                  {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "format": "uuid" },
                      "fullName": { "type": "string" },
                      "email": { "type": "string", "format": "email" },
                      "phone": { "type": "string", "nullable": true }
                    }
                  },
                  { "type": "null" }
                ]
              }
            }
          }
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "integer" },
          "message": {
            "oneOf": [
              { "type": "string" },
              {
                "type": "array",
                "items": { "type": "string" }
              }
            ]
          },
          "error": { "type": "string" }
        }
      },
      "AdminStats": {
        "type": "object",
        "properties": {
          "totalRevenue": { "type": "string", "example": "1250000.00" },
          "ordersCount": { "type": "integer" },
          "customersCount": { "type": "integer" },
          "lowStockCount": { "type": "integer" },
          "salesHistory": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "date": { "type": "string", "format": "date" },
                "amount": { "type": "string" }
              }
            }
          },
          "recentOrders": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/OrderWithItemsAndUser" }
          }
        }
      }
    }
  },
...
}