# src/orders/orders.controller.ts
import {
    Body,
    Controller,
    Get,
    Param,
    Patch,
    Post,
    Query,
    Req,
    Res,
} from '@nestjs/common';
import { ConfigService } from '@nestjs/config';
import { Request, Response } from 'express';
import { OrdersService } from './orders.service';
import { Role } from '../common/enums/role.enum';
import { CurrentUser } from '../common/decorators/current-user.decorator';
import { RequestUser } from '../common/types/request-user.type';
import { CreateOrderDto } from './dto/create-order.dto';
import { UpdateOrderStatusDto } from './dto/update-order-status.dto';
import { AdminListOrdersDto } from './dto/admin-list-orders.dto';
import { Roles } from 'src/auth/decorators/roles.decorator';
import { Public } from 'src/auth/decorators/public.decorator';
import { ensureGuestSessionId } from 'src/common/utils/guest-session.util';

@Controller()
export class OrdersController {
    constructor(
        private readonly ordersService: OrdersService,
        private readonly config: ConfigService,
    ) { }

    @Post('orders')
    @Roles(Role.CUSTOMER)
    createOrder(@CurrentUser() user: RequestUser, @Body() dto: CreateOrderDto) {
        return this.ordersService.createOrder(user.userId, dto);
    }

    @Public()
    @Post('orders/guest')
    createGuestOrder(
        @Req() request: Request,
        @Res({ passthrough: true }) response: Response,
        @Body() dto: CreateOrderDto,
    ) {
        const guestSessionId = ensureGuestSessionId(request, response, this.config);
        return this.ordersService.createGuestOrder(guestSessionId, dto);
    }

    @Get('orders/my')
    @Roles(Role.CUSTOMER)
    getMyOrders(@CurrentUser() user: RequestUser) {
        return this.ordersService.getMyOrders(user.userId);
    }

    @Get('orders/my/:id')
    @Roles(Role.CUSTOMER)
    getMyOrderById(@CurrentUser() user: RequestUser, @Param('id') id: string) {
        return this.ordersService.getMyOrderById(user.userId, id);
    }

    @Get('admin/orders')
    @Roles(Role.ADMIN)
    getAdminOrders(@Query() query: AdminListOrdersDto) {
        return this.ordersService.getAdminOrders(query);
    }

    @Get('admin/orders/:id')
    @Roles(Role.ADMIN)
    getAdminOrderById(@Param('id') id: string) {
        return this.ordersService.getAdminOrderById(id);
    }

    @Patch('admin/orders/:id/status')
    @Roles(Role.ADMIN)
    updateStatus(
        @Param('id') id: string,
        @Body() dto: UpdateOrderStatusDto,
    ) {
        return this.ordersService.updateOrderStatus(id, dto);
    }

    @Get('admin/stats')
    @Roles(Role.ADMIN)
    getAdminStats() {
        return this.ordersService.getAdminStats();
    }
}